+++ src/main/java/com/cinema/tickets_selling/service/MovieService.java
@@ -11,7 +11,7 @@ public interface MovieService extends IService<Movie> {
     Movie addMovie(Movie movie);
 
     //管理员根据影片id删除已发布影片
-    void removeMovie(Long id);
+    void removeMovie(Integer id);
 
     //管理员更新已发布影片信息
     Movie updateMovie(Movie movie);
@@ -23,7 +23,7 @@ public interface MovieService extends IService<Movie> {
     List<Movie> getUpcomingMovieList();
 
     //根据影片id获取已发布影片信息
-    Movie showMovieInfo(Long id);
+    Movie showMovieInfo(Integer id);
 
     //管理员根据影片名称查询影片相关信息
     Movie selectOneMovie(String moviename);
