+++ src/main/java/com/cinema/tickets_selling/service/OrdersService.java
@@ -12,10 +12,10 @@ public interface OrdersService extends IService<Orders> {
     Orders userAddOrder(Orders order);
 
     //用户取消订单
-    void userRemoveOrder(Long orderid);
+    void userRemoveOrder(Integer orderid);
 
     //用户根据订单id查看订单信息
-    Orders showOrderInfo(Long orderid);
+    Orders showOrderInfo(Integer orderid);
 
     //管理员根据排片id查看影片在本场次所有订单列表
     IPage<Orders> getOrdersListByScheduleId(Page p, QueryWrapper queryWrapper);
